requirements.txt
langchain
langchain-community
python-dotenv
pypdf
langchain_text_splitters
langchain_core
faiss-cpu
langchain-groq
requests
langserve
uvicorn
pandas